created a new pop-up called doctor prescription.
connected the prescription and follow-up button in accepted appointment page.

completed the function that changes the status of an appointment to completed once presciption is submitted.

completed the follow-up appointment page to send appointments to patients.

completed the recycler view on doctor homepage, enables them to click the item and take them to the appointment details.